fix: streaming ssr
